% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{amped_maintainer_and_licence}
\alias{amped_maintainer_and_licence}
\title{Show the maintainer and licence for AMPED}
\usage{
amped_maintainer_and_licence()
}
\value{
A bunch of tags for use in Shiny apps
}
\description{
Show the maintainer and licence for AMPED
}
