% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_funcs.R
\name{assessment}
\alias{assessment}
\title{assessment}
\usage{
assessment(stock, mp_params, yr, iters = 1:dim(stock$biomass)[1])
}
\arguments{
\item{stock}{The stock object}

\item{mp_params}{A named list of MP parameters (with est_sigma and est_bias elements)}

\item{yr}{The timestep that the biomass is taken from.}

\item{iters}{Numeric vector of iters. Default is all of them.}
}
\description{
Function used by \code{get_hcr_ip()} to generate input data for an assessment based HCR.
The input to the HCR is depletion (i.e. Biomass / K).
}
