% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_performance.R
\name{comparing_performance}
\alias{comparing_performance}
\title{'Comparing HCR Performance' app launcher}
\usage{
comparing_performance(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Launches the Comparing Performance Shiny app.
See the 'Information' tab in the app for more information.
Also see the package vignette (\code{vignette("comparing_performance", package="AMPLE")}) for a tutorial.
}
\examples{
\dontrun{comparing_performance()}
}
