% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Comparison plots}
\alias{Comparison plots}
\alias{plot_majuro}
\alias{quantile_plot}
\alias{myboxplot}
\alias{myradar}
\alias{pitable}
\alias{hcr_plot}
\alias{hcr_histo_plot}
\title{Plots for comparing HCR performance}
\usage{
plot_majuro(dat, hcr_choices, stock_params)

quantile_plot(dat, hcr_choices, wormdat = NULL, alpha20_80 = 0.6,
  linetype_worm = 1, colour_worm = "black", size_worm = 0.2,
  add_start_line = TRUE, time_period_lines = TRUE,
  short_term = 2016:2024, medium_term = 2025:2033,
  long_term = 2034:2042, last_plot_year = 2042,
  show_spaghetti = FALSE)

myboxplot(dat, hcr_choices, plot_type = "median_bar")

myradar(dat, hcr_choices, scaling = "scale", polysize = 2)

pitable(dat)

hcr_plot(hcr_choices, hcr_shape, hcr_points, lrp, trp, blacklinesize = 4,
  linesize = 3, pointsize = 4.2, stroke = 3)

hcr_histo_plot(hcr_choices, histodat)
}
\arguments{
\item{dat}{The data.frame with the data to be plotted.}

\item{hcr_choices}{The names of the HCRs to plot.}

\item{stock_params}{A vector of life history and stochasticy parameters.}

\item{wormdat}{Data set of the spaghetti (worms).}

\item{alpha20_80}{Alpha of the ribbons.}

\item{linetype_worm}{Line type of the spaghetti.}

\item{colour_worm}{Colour of the spaghetti.}

\item{size_worm}{Thickness of the spaghetti.}

\item{add_start_line}{Add a line to the start of the projection (TRUE / FALSE).}

\item{time_period_lines}{Add a lines to show the time periods (TRUE / FALSE).}

\item{short_term}{Year range for the short-term.}

\item{medium_term}{Year range for the medium-term.}

\item{long_term}{Year range for the long-term.}

\item{last_plot_year}{Last year to plot.}

\item{show_spaghetti}{Show the spaghetti (worms) (TRUE / FALSE).}

\item{plot_type}{Either median_bar or box.}

\item{scaling}{Either scale or rank.}

\item{polysize}{The thickness of the radar.}

\item{hcr_shape}{The shape parameters of each HCR,}

\item{hcr_points}{Optionally show the bits of the HCR that triggered (currently not used).}

\item{lrp}{The limit reference point.}

\item{trp}{The target reference point.}

\item{blacklinesize}{Size of the underlying black lines.}

\item{linesize}{Size of the coloured lines.}

\item{pointsize}{Size of the points (currently not used).}

\item{stroke}{Mmmmm, stroking.}

\item{histodat}{Data for the histograms.}
}
\value{
A ggplot2 plot object.

A ggplot2 plot object.

A ggplot2 plot object.

A ggplot2 plot object.

A data.frame to be shown as a table.

A ggplot2 plot object.

A ggplot2 plot object.
}
\description{
plot_majuro() plots a Majuro plot.

quantile_plot() plots times series of indicators for each HCR.

myboxplot() plots box plots or bar charts of the median values of the indicators for each HCR.

myradar() plots a radar chart in each time period, scaled either by the maximum or by ranking.

pitable() is not a plot but a table comparing PIs across HCRs and periods.

hcrplot() plots the shape of each HCR.

hcr_histo_plot() plots histograms of the HCR outputs in each time period.
}
