% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{create_stock}
\alias{create_stock}
\title{Make the empty reactive stock object for AMPED apps}
\usage{
create_stock()
}
\value{
A reactiveValues object with bits for the stock
}
\description{
A stock is made up of biomass, hcr_ip, hcr_op, effort and catch.
}
