% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{current_pi_table}
\alias{current_pi_table}
\title{Format and produce a table of previously calculated performance indicators}
\usage{
current_pi_table(dat, pichoice = c("pi1", "sbsbf0", "catch"))
}
\arguments{
\item{dat}{A data.frame with the 20th, 80th and 50th percentile value of each indicator.}

\item{pichoice}{A character vector of the indicator names to be included in the table}
}
\value{
A formatted data.frame
}
\description{
Format and produce a table of previously calculated performance indicators
}
