% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_funcs.R
\name{estimation_error}
\alias{estimation_error}
\title{estimation_error}
\usage{
estimation_error(input, sigma, bias)
}
\arguments{
\item{input}{A vector of the 'true' stock status}

\item{sigma}{Observation error standard deviation}

\item{bias}{Observation error bias}
}
\description{
Estimation error applied to the 'true' stock status to generate an 'observed' stock status used in the HCR.
The error is a combination of bias and lognormally distributed noise.
}
