% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_hcr_colours}
\alias{get_hcr_colours}
\title{Default palette for HCRs}
\usage{
get_hcr_colours(hcr_names, chosen_hcr_names)
}
\arguments{
\item{hcr_names}{The names of all of the HCRs}

\item{chosen_hcr_names}{The names of only the chosen HCRs}
}
\description{
Get the default palette for the HCR colours
}
