% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_funcs.R
\name{get_hcr_ip}
\alias{get_hcr_ip}
\title{Get the input to the HCR}
\usage{
get_hcr_ip(stock, mp_params, yr, ...)
}
\arguments{
\item{stock}{The stock object}

\item{mp_params}{The HCR / management procedure parameters used to evaluate the HCR (as a list).}

\item{yr}{The time step of the true stock status used to generate the HCR IP .}

\item{...}{Other arguments, including iters}
}
\description{
Run the MP analyses function to generate the input to the HCR i.e. observed stock status.
For example, estimated biomass from an assessment.
}
