% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_hcr_op}
\alias{get_hcr_op}
\title{Evaluates the HCR}
\usage{
get_hcr_op(stock, stock_params, mp_params, yr)
}
\arguments{
\item{stock}{A list with elements biomass, hcr_ip, hcr_op, effort and catch.}

\item{stock_params}{A vector of life history and stochasticy parameters.}

\item{mp_params}{A vector of management procedure parameters.}

\item{yr}{Evaluate the HCR in a particular year (timestep).}
}
\value{
A vector of outputs from the HCR.
}
\description{
Inputs to the HCRs have already been calculated and are stored in hcr_ip element of the stock.
Outputs from the HCR can be different (e.g. catch, catch multiplier, effort, effort multiplier).
Evaluates one timestep (yr), for multiple iterations.
}
