% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_projection_pis}
\alias{get_projection_pis}
\title{Get the performance indicators for the Introduction to Projections AMPED app}
\usage{
get_projection_pis(stock, stock_params, app_params, current_timestep)
}
\arguments{
\item{stock}{A list with elements biomass, hcr_ip, hcr_op, effort and catch.}

\item{stock_params}{A vector of life history and stochasticy parameters.}

\item{app_params}{A vector of application parameters.}

\item{current_timestep}{The current timestep.}
}
\value{
A filled stock object (a reactiveValues object with bits for the stock)
}
\description{
Get the performance indicators for the Introduction to Projections AMPED app
}
