% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_summaries}
\alias{get_summaries}
\title{Get the summary performance indicators}
\usage{
get_summaries(stock, stock_params, app_params, quantiles)
}
\arguments{
\item{stock}{A list with elements biomass, hcr_ip, hcr_op, effort and catch.}

\item{stock_params}{A vector of life history and stochasticy parameters.}

\item{app_params}{A vector of application parameters.}

\item{quantiles}{The quantiles to calculate the indicators over.}
}
\value{
A list of various data.frames with summaries in different formats.
}
\description{
Current indicators include SB/SBF=0, catch and the probability of being above the limit reference point.
}
