% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_hcr.R
\name{intro_hcr}
\alias{intro_hcr}
\title{Introduction to HCRs app launcher}
\usage{
intro_hcr(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Launches the introduction to HCRs Shiny app.
See the 'Information' tab in the app for more information.
Also see the package vignette (\code{vignette("intro_hcr", package="AMPLE")}) for a tutorial.
}
\examples{
\dontrun{intro_hcr()}
}
