% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuring_performance.R
\name{measuring_performance}
\alias{measuring_performance}
\title{Measuring performance app launcher}
\usage{
measuring_performance(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Launches the 'Measuring Performance' Shiny app.
See the 'Information' tab in the app for more information.
Also see the package vignette (\code{vignette("measuring_performance", package="AMPLE")}) for a tutorial.
}
\examples{
\dontrun{measuring_performance()}
}
