% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{project}
\alias{project}
\title{Projects the stock over the timesteps given and updates the biomass, HCR ip / op and catches}
\usage{
project(stock, timesteps, stock_params, mp_params, app_params,
  max_releffort = 10)
}
\arguments{
\item{stock}{A list with elements biomass, hcr_ip, hcr_op, effort and catch.}

\item{timesteps}{The timesteps to project over.}

\item{stock_params}{A vector of life history and stochasticy parameters.}

\item{mp_params}{A vector of management procedure parameters.}

\item{app_params}{A vector of application parameters.}

\item{max_releffort}{The maximum relative effort.}
}
\value{
A stock object (a reactiveValues object with bits for the stock)
}
\description{
Projects the stock over the timesteps given and updates the biomass, HCR ip / op and catches
}
