% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{reset_stock}
\alias{reset_stock}
\title{Clear out stock and refill initial period}
\usage{
reset_stock(stock, stock_params, mp_params, app_params, initial_biomass,
  nyears, niters)
}
\arguments{
\item{stock}{A list with elements biomass, hcr_ip, hcr_op, effort and catch.}

\item{stock_params}{A vector of life history and stochasticy parameters.}

\item{mp_params}{A vector of management procedure parameters.}

\item{app_params}{A vector of application parameters.}

\item{initial_biomass}{The initial biomass of the population.}

\item{nyears}{The numbers of years to set the stock up for.}

\item{niters}{The numbers of iters to set the stock up for.}
}
\value{
A filled stock object (a reactiveValues object with bits for the stock)
}
\description{
Fills the stock depending on stock status option (over, fully, underexploited).
}
