% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_module.R
\name{Stock module}
\alias{Stock module}
\alias{stockParamsSetterUI}
\alias{stockParamsSetterServer}
\alias{get_stock_params}
\title{stockParamsSetterUI}
\usage{
stockParamsSetterUI(id)

stockParamsSetterServer(id, get_stoch_params = NULL)

get_stock_params(input, biol_sigma = 0)
}
\arguments{
\item{id}{Shiny magic}

\item{get_stoch_params}{Reactive expression that accesses the stochasticity module server.}

\item{input}{List of stock parameters taken from the shiny UI (stockParamsSetterUI()).}

\item{biol_sigma}{Standard deviation of the biological variability (default = 0).}
}
\value{
A taglist

A list of stock options.
}
\description{
stockParamsSetterUI() is the interface for the stock options (e.g. life history and exploitation status).

stockParamsSetterServer() does the setting of the stock parameters in the server.

get_stock_params() Sets up default values for the stock, including year range.
It's a separate function so it can be used and tested outside of a reactive environment.
}
