% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_funcs.R
\name{threshold}
\alias{threshold}
\title{Evaluates a threshold harvest control rule}
\usage{
threshold(input, mp_params, ...)
}
\arguments{
\item{input}{A vector of the 'true' stock status}

\item{mp_params}{The HCR / management procedure parameters used to evaluate the HCR (as a list).}

\item{...}{Unused}
}
\value{
A vector of the same dimension as the input.
}
\description{
Evaluates a threshold (i.e. hockey stick) harvest control rule.
Used by the \code{hcr_op} function.
}
