% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_funcs.R
\name{get_hcr_op}
\alias{get_hcr_op}
\title{Evaluates the harvest control rule.}
\usage{
get_hcr_op(stock, mp_params, yr, iters = 1:dim(stock$biomass)[1])
}
\arguments{
\item{stock}{The stock object}

\item{mp_params}{The HCR / management procedure parameters used to evaluate the HCR (as a list).}

\item{yr}{The timestep.}

\item{iters}{A numeric vector of iters.}
}
\value{
A vector of outputs from the HCR.
}
\description{
Evaluates the harvest control rule in a single year (timestep).
}
