% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_time_periods}
\alias{get_time_periods}
\title{Get timesteps for the short-, medium- and long term time periods.}
\usage{
get_time_periods(app_params, nyears)
}
\arguments{
\item{app_params}{A vector of application parameters.}

\item{nyears}{The number of years.}
}
\value{
A list of the timesteps that make up each time period.
}
\description{
Used when calculating the performance indicators over the three time periods.
}
\examples{
app_params <- list(initial_year = 2009, last_historical_timestep = 10)
nyears <- 40
get_time_periods(app_params=app_params, nyears=nyears)
}
